# docs/api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "خالد سعد — عقد api/v1",
        "version": "1.0.0",
        "description": "مولَّدة من جدول المسارات — لا تُحرَّر يدويًّا.\n\nإصدارٌ واحد لا ثانٍ: يتطوّر api/v1 في مكانه، محروسًا ببوابة\nحدّ أدنى لإصدار التطبيق تُشحن قبل أي تغيير في العقد."
    },
    "servers": [
        {
            "url": "http://localhost/khaledsaad/public/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "object",
                        "required": [
                            "kind",
                            "code",
                            "title",
                            "message",
                            "user_action"
                        ],
                        "properties": {
                            "kind": {
                                "type": "string",
                                "enum": [
                                    "ours",
                                    "theirs",
                                    "input"
                                ]
                            },
                            "code": {
                                "type": "string"
                            },
                            "title": {
                                "type": "string"
                            },
                            "message": {
                                "type": "string"
                            },
                            "user_action": {
                                "nullable": true,
                                "description": "يجب أن يكون null حين kind = ours.",
                                "type": "object",
                                "properties": {
                                    "label": {
                                        "type": "string"
                                    },
                                    "route": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                }
                            },
                            "retry_after": {
                                "type": "integer",
                                "nullable": true
                            }
                        }
                    }
                }
            },
            "Envelope": {
                "type": "object",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object",
                        "properties": {
                            "locale": {
                                "type": "string"
                            },
                            "server_time": {
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    }
                }
            },
...
        },
        "responses": {
            "Success": {
                "description": "نجاح",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Envelope"
                        }
                    }
                }
            },
            "Error": {
                "description": "عطل مصنَّف",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        }
    },
    "paths": {
...
        "/api/v1/admin/users/plans/preview": {
            "post": {
                "operationId": "api.v1.admin.users.plans.preview",
                "summary": "api.v1.admin.users.plans.preview",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/admin/users/{user}/credits": {
            "post": {
                "operationId": "api.v1.admin.users.credits",
                "summary": "api.v1.admin.users.credits",
                "tags": [
                    "admin"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/billing": {
            "get": {
                "operationId": "api.v1.billing",
                "summary": "api.v1.billing",
                "tags": [
                    "billing"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/competitors/{competitor}/confirm": {
            "post": {
                "operationId": "api.v1.competitors.confirm",
                "summary": "api.v1.competitors.confirm",
                "tags": [
                    "competitors"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
        "/api/v1/consultations/{consultation}/confirm": {
            "post": {
                "operationId": "api.v1.consultations.confirm",
                "summary": "api.v1.consultations.confirm",
                "tags": [
                    "consultations"
                ],
                "security": [
                    {
                        "sanctum": []
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/Success"
                    },
                    "402": {
                        "$ref": "#/components/responses/Error"
                    },
                    "503": {
                        "$ref": "#/components/responses/Error"
                    }
                }
            }
        },
...
    }
}